+++ test/integration/invalid-document-import/pages/index.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line no-unused-vars
+import Document, { Main, NextScript } from 'next/document'
+
+export default () => <div>hello from index</div>
