+++ test/integration/invalid-head-import/pages/index.js
